refactor(rating): migrate scale from 1–10 integer to 0.5–5.0 decimal
- Migration: change rating column to decimal(3,1); existing values
  converted automatically (/ 2.0 up, * 2 down)
- ReadingLog: cast rating as float
- updateRating: new validation (numeric, min:0.5, max:5, multiple of 0.5);
  responds JSON when request wants JSON
- update (status): also responds JSON when request wants JSON

+++ app/Models/ReadingLog.php
@@ -26,7 +26,7 @@ class ReadingLog extends Model
      * Casts útiles.
      */
     protected $casts = [
-        'rating' => 'integer',
+        'rating' => 'float',
     ];
 
     /**
